Add MultiplatformColorScheme with Android dynamic color (Material You) support
