+++ shared/src/commonMain/composeResources/values/strings.xml
@@ -155,6 +155,8 @@
     <string name="main_settings_appearance_theme_system">System</string>
     <string name="main_settings_appearance_theme_light">Light</string>
     <string name="main_settings_appearance_theme_dark">Dark</string>
+    <string name="main_settings_appearance_dynamic_color_title">Dynamic Color</string>
+    <string name="main_settings_appearance_dynamic_color_content">Use your wallpaper colors</string>
     <plurals name="updated_duties">
         <item quantity="one">Updated %1$s duty</item>
         <item quantity="other">Updated %1$s duties</item>
