+++ shared/src/commonMain/kotlin/me/emiliomini/dutyschedule/shared/ui/theme/Theme.kt
@@ -2,69 +2,13 @@ package me.emiliomini.dutyschedule.shared.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-
-private val Red = Color(0xFFD32F2F)
-
-private val LightColorScheme = lightColorScheme(
-    primary = Red,
-    onPrimary = Color.White,
-    primaryContainer = Color(0xFFFFDAD4),
-    onPrimaryContainer = Color(0xFF410000),
-    secondary = Color(0xFF775651),
-    onSecondary = Color.White,
-    secondaryContainer = Color(0xFFFFDAD4),
-    onSecondaryContainer = Color(0xFF2C1512),
-    tertiary = Color(0xFF705C2E),
-    onTertiary = Color.White,
-    tertiaryContainer = Color(0xFFFBDFA6),
-    onTertiaryContainer = Color(0xFF251A00),
-    error = Color(0xFFBA1A1A),
-    onError = Color.White,
-    errorContainer = Color(0xFFFFDAD6),
-    onErrorContainer = Color(0xFF410002),
-    background = Color(0xFFFFFBFF),
-    onBackground = Color(0xFF201A19),
-    surface = Color(0xFFFFFBFF),
-    onSurface = Color(0xFF201A19),
-    surfaceVariant = Color(0xFFE9E9E9),
-    onSurfaceVariant = Color(0xFF534341),
-    outline = Color(0xFF857370)
-)
-
-private val DarkColorScheme = darkColorScheme(
-    primary = Color(0xFFFFB4A8),
-    onPrimary = Color(0xFF690000),
-    primaryContainer = Color(0xFF930000),
-    onPrimaryContainer = Color(0xFFFFDAD4),
-    secondary = Color(0xFFE7BDB6),
-    onSecondary = Color(0xFF442925),
-    secondaryContainer = Color(0xFF5D3F3B),
-    onSecondaryContainer = Color(0xFFFFDAD4),
-    tertiary = Color(0xFFDEC48C),
-    onTertiary = Color(0xFF3F2E04),
-    tertiaryContainer = Color(0xFF574419),
-    onTertiaryContainer = Color(0xFFFBDFA6),
-    error = Color(0xFFFFB4AB),
-    onError = Color(0xFF690005),
-    errorContainer = Color(0xFF93000A),
-    onErrorContainer = Color(0xFFFFDAD6),
-    background = Color(0xFF201A19),
-    onBackground = Color(0xFFEDE0DE),
-    surface = Color(0xFF201A19),
-    onSurface = Color(0xFFEDE0DE),
-    surfaceVariant = Color(0xFF2C2C2C),
-    onSurfaceVariant = Color(0xFFD8C2BF),
-    outline = Color(0xFFA08C8A)
-)
 
 /** themeMode: 0 = system, 1 = light, 2 = dark */
 @Composable
 fun DutyScheduleTheme(
     themeMode: Int = 0,
+    dynamicColor: Boolean = true,
     content: @Composable () -> Unit
 ) {
     val systemDark = isSystemInDarkTheme()
@@ -73,10 +17,7 @@ fun DutyScheduleTheme(
         2 -> true
         else -> systemDark
     }
-    val colorScheme = when {
-        darkTheme -> DarkColorScheme
-        else -> LightColorScheme
-    }
+    val colorScheme = platformColorScheme(darkTheme, dynamicColor)
 
     MaterialTheme(
         colorScheme = colorScheme,
